requirements.txt
numpy
pynrrd
plotly
matplotlib
chart_studio
pandas
pytest
anytree
mpi4py
scipy
matplotlib
sparse
